Basic Application setup
Initialized the application
Installed packages express and nodemon
Created the application file and folder structure
Added code to the app.js to the app that when it receives a /*
it should use the routes file
Added code for the start.js: we imported the express app
We tell the app to isten on port 3000 for incoming connections
Added code to the ./routes/index.js : the code imported the express app.
We grab the router from. Used the router to respond to any request.
Added a npm script to make nodemon start and watch our app